requirements.txt
discord.py[voice]==2.3.2
python-dotenv==1.0.0
alembic==1.11.1